Update values.yaml with better defaults
Currently the Helm Chart does not work, as the expected container port is 8080 and not 80. Furthermore the chart uses the "hostPath" as a storage class, which I do not think is preffered, as setting this value to and empty string would instead use the default storage class, which makes things easier for folks.

Signed-off-by: Nikolai Emil Damm <ned@devantler.com>